Present Consul server cert chain when using Vault secrets backend (#1251)
* Present all intermediate consul server certs

Previously, the helm chart only rendered the server leaf cert when using
the vault PKI backend. If your PKI was set up to use one or more
intermediate certs, this meant that Consul wasn't presenting the full
intermediate CA chain.

This change includes all intermediate CA certs in alongside the leaf
cert. It skips the root cert, since that's presumably already going to
be in your system's trust store.

If there's a better way to do this, I'm all ears! This is the first time
I've dealt with helm and vault templating.

* Update bats tests to match new template format

This doesn't actually test that the template is doing the _right_ thing,
but at least we can verify that the rendered template makes some sense.